Adding new package.
This is a set of tools to measure size and complexity of C files.

The original package was written by Brian Renaud. I have asked him for permission to "adopt" the package. I have written some autotools configuration files to make it easier to install, but I have not made any change to the sources (yet ;-).